requirements.txt
# Основные зависимости
telethon>=1.35.0
openai>=1.30.0
aiofiles>=24.1.0
python-dotenv>=1.0.1

# Мониторинг и производительность
psutil>=5.9.0

# Для продакшена (опционально)
uvloop>=0.19.0; sys_platform != "win32"
prometheus-client>=0.19.0

# Разработка и тестирование
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.0.0
isort>=5.12.0
mypy>=1.8.0
